+++ src/domain/vision/aruco/aruco.py
