+++ src/domain/vision/aruco/aruco_factory.py
